Where possible proxy optimized code to the legacy subpackage
Users may not know that things have improved, this allows them to benefit
in most common cases without changing their code.
